+++ common/src/main/java/net/conczin/immersive_optimization/CommonClass.java
@@ -2,6 +2,6 @@ package net.conczin.immersive_optimization;
 
 public class CommonClass {
     public static void init() {
-        Constants.LOG.info("Immersive Optimization is active!");
+        // No-op
     }
 }