+++ fabric/src/main/java/net/conczin/immersive_optimization/ImmersiveOptimization.java
@@ -1,10 +1,13 @@
 package net.conczin.immersive_optimization;
 
 import net.fabricmc.api.ModInitializer;
+import net.fabricmc.fabric.api.command.v2.CommandRegistrationCallback;
 
 public class ImmersiveOptimization implements ModInitializer {
     @Override
     public void onInitialize() {
         CommonClass.init();
+
+        CommandRegistrationCallback.EVENT.register((dispatcher, registryAccess, environment) -> Commands.register(dispatcher));
     }
 }