+++ changelog.md
@@ -1,13 +1,15 @@
 # TODO
 
 * Integrate occlusion culling
+* View distance culling
 * Per-entity profiling (mean, max, ...)
-* Server profile report commands and configuration commands
 * Client sync by chat with interaction link
   * Optional on login, otherwise just on changes by command
+* Use distant culling with shouldRenderAtSqrDistance
 
 # 0.0.1
 
 * Added basic tick scheduler and config
   * Viewport culling
   * Client linking on integrated servers
+* Server profile report and configuration commands